+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
+## 2020-02-26
+
+### Added
+
+- World bank tariff (data-store-service) pipeline
+
+
 ## 2020-02-25
 
 ### Changed